# lib/api-zod/src/generated/types/publishRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Universe Creator Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PublishRequestPayload } from './publishRequestPayload';

export interface PublishRequest {
  projectId: number;
  payload?: PublishRequestPayload;
}